Split chart and benefits list into separate components
